refactor(desktop): extract diagnostics effect hooks
